Fix codeclimate. Attempt 2
