+++ src/main/java/ru/mail/polis/service/dogm/processors/ProcessorPut.java
@@ -9,19 +9,14 @@ import ru.mail.polis.service.dogm.Bridges;
 import ru.mail.polis.service.dogm.ReplicasFraction;
 import ru.mail.polis.service.dogm.Topology;
 
-import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.logging.Logger;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.util.logging.Logger.getLogger;
 
 /**
  * Request processor for PUT method.
  */
 public class ProcessorPut extends SimpleRequestProcessor {
-    private final Logger log = getLogger("ProcessorPut");
-
     /**
      * Create a new PUT-processor.
      * @param dao DAO implementation
@@ -36,26 +31,10 @@ public class ProcessorPut extends SimpleRequestProcessor {
     public Response processEntityRequest(@NotNull final String id,
                                          @NotNull final ReplicasFraction fraction,
                                          @NotNull final Request request) {
-        int successfulResponses = 0;
-        for (final var node : topology.nodesFor(id, fraction.from)) {
-            try {
-                final Response response =
-                        topology.isMe(node)
-                                ? processEntityDirectly(id, request)
-                                : processEntityRemotely(node, request);
-                if (response.getStatus() == 201) {
-                    ++successfulResponses;
-                }
-            } catch (IOException e) {
-                log.warning(Protocol.WARN_PROCESSOR);
-            }
-        }
-
-        if (successfulResponses < fraction.ack) {
-            return new Response(Response.GATEWAY_TIMEOUT, Response.EMPTY);
-        } else {
-            return new Response(Response.CREATED, Response.EMPTY);
-        }
+        final var codeString = Response.CREATED;
+        final var codeInteger = 201;
+        return processEntityRequestOnClusterEmptyResult(
+                id, fraction, request, codeString, codeInteger);
     }
 
     @Override
